Asegurados endpoints con JWT

+++ routes/colectores.js
@@ -1,4 +1,5 @@
 const db = require('../modules/firebase');
+const jwt = require('../utils/jwt');
 let express = require('express');
 let router = express.Router();
 
@@ -15,28 +16,46 @@ router.route('/colectores').get(async function(req, res) {
 });
 
 router.route('/colectores').post(async function(req, res) {
-  const colector = {
-    nombre: req.body.nombre
-  };
-  const docRef = await db.collection('colectores').add(colector);
+  if(jwt.validateToken){
+    const colector = {
+      nombre: req.body.nombre
+    };
+    const docRef = await db.collection('colectores').add(colector);
 
-  res.json({ message: 'Colector creado', id: docRef.id });
+    res.json({ message: 'Colector creado', id: docRef.id });
+  }
+  else{
+    res.json({"message": "no autorizado"});
+  }
+
 });
 
 router.route('/colector/:id').put(async function(req, res) {
-  const colector = {
-    nombre: req.body.nombre
-  };
-  
-  await db.collection('colectores').doc(req.params.id).set(colector);
+  if(jwt.validateToken){
+    const colector = {
+      nombre: req.body.nombre
+    };
 
-  res.json({ message: 'Colector actualizado'});
+    await db.collection('colectores').doc(req.params.id).set(colector);
+
+    res.json({ message: 'Colector actualizado'});
+  }
+  else{
+    res.json({"message": "no autorizado"});
+  }
+
 });
 
 router.route('/colector/:id').delete(async function(req, res) {
-  await db.collection('colectores').doc(req.params.id).delete();
+  if(jwt.validateToken){
+    await db.collection('colectores').doc(req.params.id).delete();
 
-  res.json({ message: 'Colector eliminado' });
+    res.json({ message: 'Colector eliminado' });
+  }
+  else{
+    res.json({"message": "no autorizado"});
+  }
+
 });
 
 module.exports = router;
