+++ routes/especies.js
@@ -1,8 +1,10 @@
 const db = require('../modules/firebase');
 const express = require('express');
+const jwt = require('../utils/jwt');
 const router = express.Router();
 
 router.route('/especies').get(async function(req, res) {
+
   const especiesSnapshot = await db.collection('especies').get();
   const especies = [];
   especiesSnapshot.forEach(especie => {
@@ -15,28 +17,46 @@ router.route('/especies').get(async function(req, res) {
 });
 
 router.route('/especies').post(async function(req, res) {
-  const especie = {
-    nombre: req.body.nombre
-  };
-  const docRef = await db.collection('especies').add(especie);
+  if(jwt.validateToken){
+    const especie = {
+      nombre: req.body.nombre
+    };
+    const docRef = await db.collection('especies').add(especie);
 
-  res.json({ message: 'Especie creada', id: docRef.id });
+    res.json({ message: 'Especie creada', id: docRef.id });
+  }
+  else{
+    res.json({"message": "no autorizado"});
+  }
+
 });
 
 router.route('/especie/:id').put(async function(req, res) {
-  const especie = {
-    nombre: req.body.nombre
-  };
+  if(jwt.validateToken){
+    const especie = {
+      nombre: req.body.nombre
+    };
 
-  await db.collection('especies').doc(req.params.id).set(especie);
+    await db.collection('especies').doc(req.params.id).set(especie);
 
-  res.json({ message: 'Especie actualizada' });
+    res.json({ message: 'Especie actualizada' });
+  }
+  else{
+    res.json({"message": "no autorizado"});
+  }
+
 });
 
 router.route('/especie/:id').delete(async function(req, res) {
-  await db.collection('especies').doc(req.params.id).delete();
+  if(jwt.validateToken){
+    await db.collection('especies').doc(req.params.id).delete();
 
-  res.json({ message: 'Especie eliminada' });
+    res.json({ message: 'Especie eliminada' });
+  }
+  else{
+    res.json({"message": "no autorizado"});
+  }
+
 });
 
 module.exports = router;
