+++ routes/lugares.js
@@ -1,5 +1,6 @@
 const db = require('../modules/firebase');
 const express = require('express');
+const jwt = require('../utils/jwt');
 const router = express.Router();
 
 router.route('/lugares').get(async function(req, res) {
@@ -18,32 +19,50 @@ router.route('/lugares').get(async function(req, res) {
 });
 
 router.route('/lugares').post(async function(req, res) {
-  const lugar = {
-    pais: req.body.pais,
-    departamento: req.body.departamento,
-    municipio: req.body.municipio,
-    localidad: req.body.localidad
-  }
-  const docRef = await db.collection('lugares').add(lugar);
+  if(jwt.validateToken){
+    const lugar = {
+      pais: req.body.pais,
+      departamento: req.body.departamento,
+      municipio: req.body.municipio,
+      localidad: req.body.localidad
+    }
+    const docRef = await db.collection('lugares').add(lugar);
 
-  res.json({ message: 'Lugar creado', id: docRef.id });
+    res.json({ message: 'Lugar creado', id: docRef.id });
+  }
+  else{
+    res.json({"message": "no autorizado"});
+  }
+
 });
 
 router.route('/lugar/:id').put(async function(req, res) {
-  const lugar = {};
-  for (prop in req.body) {
-    lugar[prop] = req.body[prop];
+  if(jwt.validateToken){
+    const lugar = {};
+    for (prop in req.body) {
+      lugar[prop] = req.body[prop];
+    }
+
+    await db.collection('lugares').doc(req.params.id).set(lugar, { merge: true });
+
+    res.json({ message: 'Lugar actualizado' });
+  }
+  else{
+    res.json({"message": "no autorizado"});
   }
 
-  await db.collection('lugares').doc(req.params.id).set(lugar, { merge: true });
-
-  res.json({ message: 'Lugar actualizado' });
 });
 
 router.route('/lugar/:id').delete(async function(req, res) {
-  await db.collection('lugares').doc(req.params.id).delete();
+  if(jwt.validateToken){
+    await db.collection('lugares').doc(req.params.id).delete();
 
-  res.json({ message: 'Lugar eliminado' });
+    res.json({ message: 'Lugar eliminado' });
+  }
+  else{
+    res.json({"message": "no autorizado"});
+  }
+
 });
 
 module.exports = router;
